% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_hmisc_summary.R
\name{display_hmisc_summary}
\alias{display_hmisc_summary}
\title{Display table}
\usage{
display_hmisc_summary(table1, npct = "both", ...)
}
\arguments{
\item{table1}{a table returned by summaryM etc in Hmisc}

\item{npct}{specifies which counts are to be printed to the right of percentages. You can specify "both" to print both numerator and denominator as a fraction, "denominator", "slash" to typeset horizontally using a forward slash, or "none".}
}
\description{
Display table
}
