% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_rms_summary.R
\name{display_rms_summary}
\alias{display_rms_summary}
\title{Display summary from a model generated by rms}
\usage{
display_rms_summary(m, width = 150)
}
\arguments{
\item{width}{the width of the table in pixels}

\item{s}{a model returned by rms}
}
\description{
Display summary from a model generated by rms
}
