% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_df.r
\name{display_df}
\alias{display_df}
\title{Displays a formatted table}
\usage{
display_df(df, width = 100)
}
\arguments{
\item{df}{a data.frame}

\item{width}{the width of the columns in pixels}
}
\description{
Displays a formatted table
}
\examples{
\dontrun{
df <- data.frame(a=c(1,2,3),b=c(4,5,6))
display_df(df)
}
}
