% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_hmisc_summary.R
\name{display_hmisc_summary}
\alias{display_hmisc_summary}
\title{Display table}
\usage{
display_hmisc_summary(table1)
}
\arguments{
\item{table1}{a table returned by summaryM etc in Hmisc}
}
\description{
Display table
}
